loadPrivileges-Methode in den SqlProcedureController schieben
